Trait pallet_root_of_trust::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Currency: Currency<Self::AccountId>;
    type CertificateId: Member + Parameter + MaybeSerializeDeserialize + Debug + MaybeDisplay + Ord + Default;
    type SlotBookingCost: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type SlotRenewingCost: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>;
    type SlotValidity: Get<Self::BlockNumber>;
    type FundsCollector: OnUnbalanced<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::NegativeImbalance>;
    type WeightInfo: WeightInfo;
}

Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type Currency: Currency<Self::AccountId>

The currency used to represent the voting power

type CertificateId: Member + Parameter + MaybeSerializeDeserialize + Debug + MaybeDisplay + Ord + Default

How a certificate public key is represented, typically AccountId

type SlotBookingCost: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

How much a new root certificate costs

type SlotRenewingCost: Get<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::Balance>

How much renewing a root certificate costs

type SlotValidity: Get<Self::BlockNumber>

How long a certificate is considered valid

type FundsCollector: OnUnbalanced<<<Self as Config>::Currency as Currency<<Self as Config>::AccountId>>::NegativeImbalance>

The module receiving funds paid by depositors, typically a company reserve

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.